Move authenticate decorator to utils [#161897188]
